+++ projects/ngx-tools/src/lib/future-builder/directives/when-empty-error.directive.ts
@@ -2,9 +2,10 @@ import {ChangeDetectorRef, Directive, Input, OnDestroy, TemplateRef, ViewContain
 import {Subject, Subscription, switchMap} from "rxjs";
 import {map} from "rxjs/operators";
 import {FutureSwitch} from "../models/future-switch.model";
+import {BaseFutureRender} from "../models/base-future.render";
 
 @Directive({selector: '[whenEmptyError], [whenEmptyError][whenEmptyErrorFrom]'})
-export class WhenEmptyErrorDirective<T> implements OnDestroy {
+export class WhenEmptyErrorDirective<T> extends BaseFutureRender<TemplateContext<T>> implements OnDestroy {
 
   sub: Subscription;
   states$ = new Subject<FutureSwitch<T>>();
@@ -21,18 +22,16 @@ export class WhenEmptyErrorDirective<T> implements OnDestroy {
   }
 
   constructor(
-    private templateRef: TemplateRef<TemplateContext<T>>,
-    private viewContainer: ViewContainerRef,
-    private changes: ChangeDetectorRef
+    templateRef: TemplateRef<TemplateContext<T>>,
+    viewContainer: ViewContainerRef,
+    changes: ChangeDetectorRef
   ) {
+    super(templateRef, viewContainer, changes);
+
     this.sub = this.states$.pipe(
       switchMap(x => x.emptyError$),
       map(x => x ? {$implicit: x.error} as TemplateContext<T> : undefined)
-    ).subscribe(c => {
-      if (c) this.viewContainer.createEmbeddedView(this.templateRef, c);
-      else this.viewContainer.clear();
-      this.changes.detectChanges();
-    });
+    ).subscribe(c => this.updateView(c));
   }
 
   ngOnDestroy() {