+++ projects/ngx-tools/src/lib/future-builder/directives/when-loading.directive.ts
@@ -3,9 +3,10 @@ import {Subject, Subscription, switchMap} from "rxjs";
 import {map} from "rxjs/operators";
 import {FutureLoading} from "@consensus-labs/rxjs-tools";
 import {FutureSwitch} from "../models/future-switch.model";
+import {BaseFutureRender} from "../models/base-future.render";
 
 @Directive({selector: '[whenLoading]'})
-export class WhenLoadingDirective<T> implements OnDestroy {
+export class WhenLoadingDirective<T> extends BaseFutureRender<TemplateContext<T>> implements OnDestroy {
 
   sub: Subscription;
   states$ = new Subject<FutureSwitch<T>>();
@@ -17,21 +18,19 @@ export class WhenLoadingDirective<T> implements OnDestroy {
   }
 
   constructor(
-    private templateRef: TemplateRef<TemplateContext<T>>,
-    private viewContainer: ViewContainerRef,
-    private changes: ChangeDetectorRef
+    templateRef: TemplateRef<TemplateContext<T>>,
+    viewContainer: ViewContainerRef,
+    changes: ChangeDetectorRef
   ) {
+    super(templateRef, viewContainer, changes);
+
     this.sub = this.states$.pipe(
       switchMap(x => x.loading$),
       map(x => x ? {
         data: x instanceof FutureLoading ? x.value : undefined,
         loading: x instanceof FutureLoading
       } as TemplateContext<T> : undefined)
-    ).subscribe(c => {
-      if (c) this.viewContainer.createEmbeddedView(this.templateRef, c);
-      else this.viewContainer.clear();
-      this.changes.detectChanges();
-    });
+    ).subscribe(c => this.updateView(c));
   }
 
   ngOnDestroy() {